Move verification tests to idaes-compatibility (#1068)
* Reworking some verification tests into examples

* Moving another verification test to compatibility repo
